requirements.txt
pytest
pytest-mock
parameterized
flake8

# for parallel execution
pytest-xdist

# ui testing
selenium

# rest api testing
requests

# load testing
sanic
flask
locust

# for android native and Windows
Appium-Python-Client>=0.28
